restclient: Added pkg/api to store req/resp structs
Also moved /restclient to /pkg/restclient

Updates: #265
