+++ e2e/peer_ops_test.go
@@ -1,11 +1,13 @@
 package e2e
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"strings"
 	"testing"
 
+	"github.com/gluster/glusterd2/pkg/api"
 	"github.com/stretchr/testify/require"
 )
 
@@ -25,8 +27,13 @@ func TestAddRemovePeer(t *testing.T) {
 	r.True(g2.IsRunning())
 
 	// add peer: ask g1 to add g2 as peer
-	reqBody := strings.NewReader(fmt.Sprintf(`{"addresses": ["%s"]}`, g2.PeerAddress))
-	resp, err := http.Post("http://"+g1.ClientAddress+"/v1/peers", "application/json", reqBody)
+	peerAddReq := api.PeerAddReq{
+		Addresses: []string{g2.PeerAddress},
+	}
+	reqBody, err := json.Marshal(peerAddReq)
+	r.Nil(err)
+
+	resp, err := http.Post("http://"+g1.ClientAddress+"/v1/peers", "application/json", strings.NewReader(string(reqBody)))
 	r.Nil(err)
 	defer resp.Body.Close()
 	r.Equal(resp.StatusCode, 201)
