+++ pkg/restclient/errors.go
@@ -1,9 +1,16 @@
 package restclient
 
+import (
+	"fmt"
+)
+
+// APIUnsupportedError is custom error struct for API 404 errors
 type APIUnsupportedError struct {
 	msg string
 }
 
+// APIUnexpectedStatusCodeError is custom error when expected
+// status code does not match with return status code
 type APIUnexpectedStatusCodeError struct {
 	msg      string
 	expected int
@@ -11,7 +18,7 @@ type APIUnexpectedStatusCodeError struct {
 	resp     string
 }
 
-func NewAPIUnsupportedError() error {
+func raiseAPIUnsupportedError() error {
 	return &APIUnsupportedError{"Unsupported API"}
 }
 
@@ -19,7 +26,7 @@ func (e *APIUnsupportedError) Error() string {
 	return e.msg
 }
 
-func NewAPIUnexpectedStatusCodeError(expected int, actual int, resp string) error {
+func raiseAPIUnexpectedStatusCodeError(expected int, actual int, resp string) error {
 	return &APIUnexpectedStatusCodeError{"Unexpected Status Code", expected, actual, resp}
 }
 
@@ -27,4 +34,3 @@ func (e *APIUnexpectedStatusCodeError) Error() string {
 	msg := fmt.Sprintf("%s Expected: %d Actual: %d (%s)", e.msg, e.expected, e.actual, e.resp)
 	return msg
 }
-
