+++ pkg/restclient/examples/main.go
